sys-requirements.txt
python3-venv
python3-pip
python3-lgpio
python3-gpiozero
python3-pigpio
python3-serial